Dropping (attempted) support for all lowercase `@defaultvalue`
The official documentation says that `@defaultvalue` should be all lowercase.

Implementing support for this was more difficult than I expected. I'm droping this as being out of scope for this PR.

I have no intention of using the all lowercase `@defaultvalue` tag so I don't care about trying to support it.