fix: default tag type bug (#160)
* Getting the debugger to work for me

* #159 - Adding tests for expected default values

* FIX: `@default []` and `@default {}` turns into `@default` #159

* #159 Updating the `@default` tests

* Don't need to default to an array

This was outdated code from an earlier idea I had

* Adding `@defaultValue` tests

* Adding tests for non empty defaults

* Attempt at using the fix method mentioned in comment-parser issue 155

https://github.com/syavorsky/comment-parser/issues/155

* Revert "Attempt at using the fix method mentioned in comment-parser issue 155"

This reverts commit 52800c571a037c40848b819072f00db77234cb87.

* Update to the `@default` tests (Warning: tests are failing)

* Exporting util funcitons directly rather than at the end

* Moving the `@default` tag value into the type property

* Updating expectations for `@default` tests

* Almost working, only 1 failing test left to fix

* Failed attempt at getting `@default {{ obj: 'val' }}` to work

* Dropping attempted support for curly bracket boilerplate `@default` value

It is being too troublesome to implement.
It also isn't part of the official jsDoc spec.
I think people would most likely either use values directly or use the JS doc official syntax so I don't see much reason in trying to support it.

* simplifying squareBracketBoilerplateMatch

I'm not trying to match against both square and curly boilerplates so I only need the one boiler plate type to match against

* Adding empty array/object in square bracket boilerplate tests

* Revert "Exporting util funcitons directly rather than at the end"

This reverts commit cc89717db896274cbb10c3a0f273992aac4aadde.

* Adding `@defaultvalue` (all lowercase) as a valid tag type

* Removing support for `@default [value]` syntax

That was a misunderstanding. The documentation means that the value is optional. It doesn't want you to put litteral square brackets in your code.

* Updating tests

* Making use of new `TAGS_DEFAULT` role for less repitition

* Dropping (attempted) support for all lowercase `@defaultvalue`

The official documentation says that `@defaultvalue` should be all lowercase.

Implementing support for this was more difficult than I expected. I'm droping this as being out of scope for this PR.

I have no intention of using the all lowercase `@defaultvalue` tag so I don't care about trying to support it.

* order fix in TAGS_PEV_FORMATE_DESCRIPTION

Co-authored-by: Daniel Tonon <dan503.npm@gmail.com>